rename klogtologrus to klogtolog
