+++ klogtolog/test/klog-powered-lib/lib_methods.go
