+++ klogtolog/test/service/some_service.go
@@ -3,8 +3,8 @@ package service
 import (
 	log "github.com/deckhouse/deckhouse/pkg/log"
 
-	_ "github.com/flant/kube-client/klogtologrus" // plug in JSON logs adapter
-	klog_powered_lib "github.com/flant/kube-client/klogtologrus/test/klog-powered-lib"
+	_ "github.com/flant/kube-client/klogtolog" // plug in JSON logs adapter
+	klog_powered_lib "github.com/flant/kube-client/klogtolog/test/klog-powered-lib"
 )
 
 func DoWithCallToKlogPoweredLib() {
